+++ contract/contracts/XtaxPlanet.sol
@@ -216,8 +216,8 @@ contract XtaxPlanet is ERC721, Ownable {
 
     function removeFromRecentCreations(uint256 tokenId) internal {
         for(uint8 i = 0; i < NUM_RECENT_CREATIONS; i++) {
-            if(CircularBuffer.read(ownerToRecentCreations[msg.sender], i) == tokenId) {
-                CircularBuffer.erase(ownerToRecentCreations[msg.sender], i);
+            if(CircularBuffer.read(ownerToRecentCreations[msg.sender], int8(i)) == tokenId) {
+                CircularBuffer.erase(ownerToRecentCreations[msg.sender], int8(i));
                 break;
             }
         }
@@ -227,7 +227,7 @@ contract XtaxPlanet is ERC721, Ownable {
         PlanetInfo[] memory planetInfos = new PlanetInfo[](NUM_RECENT_CREATIONS);
 
         for(uint8 i = 0; i < NUM_RECENT_CREATIONS; i++) {
-            uint256 tokenId = CircularBuffer.read(ownerToRecentCreations[owner], i);
+            uint256 tokenId = CircularBuffer.read(ownerToRecentCreations[owner], int8(i));
             planetInfos[i] = tokenIdToInfo[tokenId];
         }
